+++ src/test/scala/io/bartholomews/discogs4s/samples/PersonalTokenReadme.scala
@@ -2,16 +2,19 @@ package io.bartholomews.discogs4s.samples
 
 object PersonalTokenReadme {
   import io.bartholomews.discogs4s.DiscogsClient
-  import io.bartholomews.discogs4s.entities.{SimpleUser, Username}
+  import io.bartholomews.discogs4s.entities.UserIdentity
   import io.bartholomews.fsclient.core.http.SttpResponses.SttpResponse
+  import io.bartholomews.fsclient.core.oauth.OAuthSigner
   import io.circe
   import sttp.client.{HttpURLConnectionBackend, Identity, NothingT, SttpBackend}
 
   type F[X] = Identity[X]
   implicit val backend: SttpBackend[F, Nothing, NothingT] = HttpURLConnectionBackend()
 
-  private val client = DiscogsClient.personalFromConfig
+  private val discogs = DiscogsClient.personalFromConfig
 
-  val response: F[SttpResponse[circe.Error, SimpleUser]] =
-    client.users.getSimpleUserProfile(Username("_.bartholomews"))
+  implicit val personalToken: OAuthSigner = discogs.client.signer
+
+  // You can make authenticated (for your user only) calls with the implicit signer
+  val response: F[SttpResponse[circe.Error, UserIdentity]] = discogs.users.me
 }